// engine/enums/field-zone.ts
export enum FieldZone {
  DefenseLeft = "DefenseLeft",
  DefenseCenter = "DefenseCenter",
  DefenseRight = "DefenseRight",
  MidfieldLeft = "MidfieldLeft",
  MidfieldCenter = "MidfieldCenter",
  MidfieldRight = "MidfieldRight",
  AttackLeft = "AttackLeft",
  AttackCenter = "AttackCenter",
  AttackRight = "AttackRight",
}